requirements.txt
# AI Employee Dependencies
# Install with: pip install -r requirements.txt

# File System Watcher
watchdog>=3.0.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0

# WhatsApp (Playwright)
playwright>=1.40.0

# LinkedIn (Playwright)
# (uses same playwright package)

# Scheduler
schedule>=1.2.0

# Environment Variables
python-dotenv>=1.0.0
